% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icams_replacements.R
\name{TransformCatalog}
\alias{TransformCatalog}
\title{Transform catalog between different types}
\usage{
TransformCatalog(catalog, target.catalog.type)
}
\arguments{
\item{catalog}{Input catalog}

\item{target.catalog.type}{Target catalog type}
}
\value{
Transformed catalog
}
\description{
Transform catalog between different types
}
\keyword{internal}
