% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icams_replacements.R
\name{WriteCatalog}
\alias{WriteCatalog}
\title{Write a catalog to CSV file}
\usage{
WriteCatalog(catalog, file)
}
\arguments{
\item{catalog}{The catalog to write}

\item{file}{The file path to write to}
}
\description{
Write a catalog to CSV file
}
\keyword{internal}
